Created Billing Module with DC's and Invoice Sent Automatically (#10)

+++ src/main/java/com/kalibyte/foundry/order/dto/response/OrderItemResponse.java
@@ -15,7 +15,7 @@ public class OrderItemResponse {
     private UUID id;
     private String productName;
     private String metalType;
-    private BigDecimal quantity;
+    private int quantity;
     private BigDecimal unitPrice;
     private BigDecimal totalPrice;
 }
